mcp_servers/kuka_knowledge/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kuka-knowledge"
version = "0.1.0"
description = "MCP server exposing the kuka_dataset/normalized/ knowledge base via semantic + keyword search."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "The Way Automation LLC" }]
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
embeddings = [
    "numpy>=1.26",
]

[tool.hatch.build.targets.wheel]
packages = ["kuka_knowledge"]
